--- a/cpp/src/plasma/CMakeLists.txt
+++ b/cpp/src/plasma/CMakeLists.txt
@@ -77,6 +77,7 @@ set(PLASMA_SRCS
   io.cc
   malloc.cc
   plasma.cc
+  plasma_queue.cc
   protocol.cc
   thirdparty/ae/ae.c
   thirdparty/xxhash.cc)
@@ -143,6 +144,7 @@ install(FILES
   compat.h
   client.h
   events.h
+  plasma_queue.h
   test-util.h
   DESTINATION "${CMAKE_INSTALL_INCLUDEDIR}/plasma")