--- a/cpp/src/plasma/format/common.fbs
+++ b/cpp/src/plasma/format/common.fbs
@@ -36,4 +36,5 @@ table ObjectInfo {
   digest: string;
   // Specifies if this object was deleted or added.
   is_deletion: bool;
+  type: int;
 }