This PR is intended to share the plasma queue and performance testing code in arrow, so that people can potential make modifications and compare performance.

The performance testing code is in client_tests. 

The following test cases map to the performance testing section of [plasma queue proposal](https://docs.google.com/document/d/11MiOns63QTK_Apxy6VLj_zhmO1wCRZ06pB2UZoUFKOY/edit#).

zhijunfu@localhost:~/Code/arrow.ant-tech-alliance.queue-perf/cpp/build$ ./release/client_tests
[ RUN      ] TestPlasmaStore.BatchQueueItemPerfTest        <--- _Current plasma queue, case [4]_
[start] Plasma queue performance test           
PushQueueItem started
PushQueueItem takes 1.06687 seconds
GetQueueItem started
GetQueueItem takes 1.09613 seconds
[done] Plasma queue performance test
[       OK ] TestPlasmaStore.BatchQueueItemPerfTest (2523 ms)
[ RUN      ] TestPlasmaStore.BatchQueueItemPerfTest2      <--- __Let store manage queue buffer, case_ [2]_
[start] Plasma queue performance test 2       
PushQueueItem started
PushQueueItem takes 3.59372 seconds
[       OK ] TestPlasmaStore.BatchQueueItemPerfTest2 (3951 ms)
[ RUN      ] TestPlasmaStore.BatchCreateObjectPerfTest     <--- _batch 1000 objects, case [5]_
Generate object ids started
Generate object ids takes 0.189514 seconds
[start] Plasma object performance test    <--- 
Push objects started
Push object takes 1.67594 seconds
Get object started
Get object takes 1.05948 seconds
[done] Plasma object performance test
[       OK ] TestPlasmaStore.BatchCreateObjectPerfTest (4894 ms)
[ RUN      ] TestPlasmaStore.BatchObjectPerfTest        <--- _using objects, case [1]_
[start] Plasma object performance test
Push objects started
Push object takes 5.054 seconds
Get object started
Get object takes 6.02782 seconds
[done] Plasma object performance test